backend/requirements.txt
fastapi
uvicorn

# for databases
sqlalchemy
psycopg2

# for environment variables
python-dotenv
